Added memo function DateString
